test: merge per-module tests into existing files per review
Fold the new cover/datetime/remote/config_flow tests into their existing
per-module test files and remove the parallel *_unit/_extended files, per
maintainer review. test_entity.py and test_cloud.py are kept as new files
since entity.py and cloud.py had no existing test file to merge into.
